First commit of files for this project. The session files used here are referenced from a youtube video that I produced https://youtu.be/CHzfHG3ah1g
Here is the first commit of files.

This includes the Octatrack Project files and the OP-Z MIDI config file.

In order to use, you will need to create a set or choose an existing set folder on your Octactrack and then copy the project folder over to the set directory on the CF Media card. If you need further info on how to do this, check Elektron's manual or website.

To use the OP-Z MIDI config file, you connect the OP-Z as a disk drive, then overwrite the midi.json file on the OP-Z.

Remember to back up your work before making any changes and have fun.